+++ src/middleware/account-language-change.ts
@@ -1,5 +1,5 @@
 /**
- * @file user-change-language.ts - middleware for user change language
+ * @file account-language-change.ts - middleware for user change language
  * @author cmc
  * @since v0.0.1
  */
