+++ src/middleware/index.ts
@@ -1,18 +1,19 @@
 import { Application } from '../declarations'
 // import { HookContext } from '@feathersjs/feathers'
-import confirmEmail from './confirm-email'
-import createUser from './create-user'
-import resetPassword from './reset-password'
-import setNewPassword from './set-new-password'
-import nameTaken from './name-taken'
-import emailTaken from './email-taken'
-import editors from './editors'
-import userRole from './user-role'
-import userChangeLanguage from './user-change-language'
+import confirmEmail from './account-email-confirm'
+import createUser from './account-create'
+import resetPassword from './account-password-reset'
+import setNewPassword from './account-password-set'
+import nameTaken from './account-name-taken'
+import emailTaken from './account-email-taken'
+import editors from './editors-count'
+import userRole from './account-role'
+import userChangeLanguage from './account-language-change'
+import accountChangeRole from './account-role-change'
 // Don't remove this comment. It's needed to format import lines nicely.
 
 export default function (app: Application): void {
-  app.use('/accounts/:id/change-language/', userChangeLanguage(app))
+  app.use('/accounts/:id/change-language', userChangeLanguage(app))
   app.use('/accounts/:id/role', userRole(app))
   app.use('/accounts/confirm', confirmEmail(app))
   app.use('/accounts/create', createUser(app))
